Preliminary form support
Implement Form.elm to get some manual input from the user. When a form
is shown normal preventDefault is stopped via ports. This needs to be
done so the user can type into the input boxes without triggering any
buttons. Some more work is needed to make the experience smooth.