Various refactorings related to Select and Highlight
Moved highlight and selection information from workspace to tool.
Removed Tool dependency on Visual. Implemented Interaction type which
is used for both selection and highlight and allows more fine grained
control on how the selections and highlights should be visualized (for
example line segments, nets, and  primary and secondary highlights)